- Relocated hierarchical-namespace/delete_empty_folder.py to be under storagecontrol/. - Fixed bad indent. - Tested script.
